--- crossterm_cursor/src/cursor/test.rs
@@ -1,92 +1,0 @@
-#![allow(unused_must_use)]
-
-use super::AnsiCursor;
-use super::ITerminalCursor;
-
-/* ======================== WinApi =========================== */
-#[cfg(windows)]
-mod winapi_tests {
-    use super::super::WinApiCursor;
-    use super::*;
-
-    #[test]
-    fn goto_winapi() {
-        let cursor = WinApiCursor::new();
-
-        cursor.goto(5, 5);
-        let (x, y) = cursor.pos();
-
-        assert_eq!(x, 5);
-        assert_eq!(y, 5);
-    }
-
-    #[test]
-    fn reset_safe_winapi() {
-        let cursor = WinApiCursor::new();
-        let (x, y) = cursor.pos();
-
-        cursor.save_position();
-        cursor.goto(5, 5);
-        cursor.reset_position();
-
-        let (x_saved, y_saved) = cursor.pos();
-
-        assert_eq!(x, x_saved);
-        assert_eq!(y, y_saved);
-    }
-}
-
-/* ======================== ANSI =========================== */
-// TODO - Test is ingored, because it's stalled on Travis CI
-#[test]
-#[ignore]
-fn reset_safe_ansi() {
-    if try_enable_ansi() {
-        let cursor = AnsiCursor::new();
-        let (x, y) = cursor.pos();
-
-        cursor.save_position();
-        cursor.goto(5, 5);
-        cursor.reset_position();
-
-        let (x_saved, y_saved) = cursor.pos();
-
-        assert_eq!(x, x_saved);
-        assert_eq!(y, y_saved);
-    }
-}
-
-// TODO - Test is ingored, because it's stalled on Travis CI
-#[test]
-#[ignore]
-fn goto_ansi() {
-    if try_enable_ansi() {
-        let cursor = AnsiCursor::new();
-        let (x_saved, y_saved) = cursor.pos();
-
-        cursor.goto(5, 5);
-        let (x, y) = cursor.pos();
-
-        cursor.goto(x_saved, y_saved);
-
-        assert_eq!(x, 5);
-        assert_eq!(y, 5);
-    }
-}
-
-fn try_enable_ansi() -> bool {
-    #[cfg(windows)]
-    {
-        if cfg!(target_os = "windows") {
-            use crossterm_utils::sys::winapi::ansi::set_virtual_terminal_processing;
-
-            // if it is not listed we should try with WinApi to check if we do support ANSI-codes.
-            match set_virtual_terminal_processing(true) {
-                Ok(_) => return true,
-                Err(_) => return false,
-            }
-        }
-    }
-
-    true
-}
